+++ x/foundation/keeper/internal/hooks.go
@@ -1,0 +1,83 @@
+package internal
+
+import (
+	"context"
+
+	"cosmossdk.io/math"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+)
+
+var _ stakingtypes.StakingHooks = Hooks{}
+
+// Wrapper struct
+type Hooks struct {
+	k Keeper
+}
+
+// Create new foundation hooks
+func (k Keeper) Hooks() Hooks { return Hooks{k} }
+
+func (h Hooks) AfterDelegationModified(_ context.Context, _ sdk.AccAddress, _ sdk.ValAddress) error {
+	return nil
+}
+
+func (h Hooks) AfterUnbondingInitiated(_ context.Context, _ uint64) error {
+	return nil
+}
+
+func (h Hooks) AfterValidatorBeginUnbonding(_ context.Context, _ sdk.ConsAddress, _ sdk.ValAddress) error {
+	return nil
+}
+
+func (h Hooks) AfterValidatorBonded(_ context.Context, _ sdk.ConsAddress, _ sdk.ValAddress) error {
+	return nil
+}
+
+func (h Hooks) AfterValidatorCreated(goCtx context.Context, valAddr sdk.ValAddress) error {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	grantee := sdk.AccAddress(valAddr)
+
+	valAddrStr, err := h.k.cdc.InterfaceRegistry().SigningContext().ValidatorAddressCodec().BytesToString(valAddr)
+	if err != nil {
+		return err
+	}
+
+	// This hook isn't run by any msgs other than MsgCreateValidator,
+	// so msg is able to be reconstruct without using ctx.TxBytes()
+	msg := stakingtypes.MsgCreateValidator{
+		ValidatorAddress: valAddrStr,
+	}
+
+	if err := h.k.Accept(ctx, grantee, &msg); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (h Hooks) AfterValidatorRemoved(_ context.Context, _ sdk.ConsAddress, _ sdk.ValAddress) error {
+	return nil
+}
+
+func (h Hooks) BeforeDelegationCreated(_ context.Context, _ sdk.AccAddress, _ sdk.ValAddress) error {
+	return nil
+}
+
+func (h Hooks) BeforeDelegationRemoved(_ context.Context, _ sdk.AccAddress, _ sdk.ValAddress) error {
+	return nil
+}
+
+func (h Hooks) BeforeDelegationSharesModified(_ context.Context, _ sdk.AccAddress, _ sdk.ValAddress) error {
+	return nil
+}
+
+func (h Hooks) BeforeValidatorModified(_ context.Context, _ sdk.ValAddress) error {
+	return nil
+}
+
+func (h Hooks) BeforeValidatorSlashed(_ context.Context, _ sdk.ValAddress, _ math.LegacyDec) error {
+	return nil
+}