+++ x/foundation/module/module.go
@@ -22,6 +22,7 @@ import (
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 	paramstypes "github.com/cosmos/cosmos-sdk/x/params/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
 	modulev1 "github.com/Finschia/finschia-sdk/api/lbm/foundation/module/v1"
 	"github.com/Finschia/finschia-sdk/x/foundation"
@@ -219,6 +220,7 @@ type FoundationOutputs struct {
 
 	Keeper keeper.Keeper
 	Module appmodule.AppModule
+	Hooks  stakingtypes.StakingHooksWrapper
 }
 
 func ProvideModule(in FoundationInputs) FoundationOutputs {
@@ -250,6 +252,7 @@ func ProvideModule(in FoundationInputs) FoundationOutputs {
 	return FoundationOutputs{
 		Keeper: k,
 		Module: m,
+		Hooks:  stakingtypes.StakingHooksWrapper{StakingHooks: k.Hooks()},
 	}
 }